requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.2
face-recognition==1.3.0
boto3==1.34.10
botocore==1.34.10
python-dateutil==2.8.2
pytz==2023.3
Werkzeug==3.0.1
